requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
diff-match-patch==20121119
Django==1.11.7
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.5.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.10
django-simple-captcha==0.5.4
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.6
jdcal==1.3
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
pkg-resources==0.0.0
PyMySQL==0.7.11
PyMySQL3==0.5
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
